lib/pyproject.toml
[project]
name = "sandbox-common"
version = "0.1.0"
description = "Shared library for agent-sandboxes applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e2b>=2.6.4",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
